MB-55912: Temporarily disable history for default collection
kv_engine is in the middle of merging forwards the CDC work andd they
originally blocked history for the default collection and appear to have
other bugs. Temporarily disable history for the default collection to
make Elixir testing easier. We can revert this change in the future.

Change-Id: I635656fbb5346de4a9de1d630492ccb5a987998d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188088
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>