MB-55922: Accept reference to old node in janitor_agent.
Topology information passed in set_vbucket_state calls are persisted in
KV (MB-33389). Therefore when a bucket is paused and resumed on a new
cluster, the first stats call to memcached returns a topology for the
vbucket with a reference to the old nodes.

Fix decode_topology/1 to accept any old nodes.

Change-Id: Id202bd305322ef3e332d64567aea0aa33914f3d5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188107
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>