MB-55955: Set JAVA_HOME for api_tests
When the JAVA_HOME environment variable is not set but java can be
found in the PATH, the LdapServer() call will hang indefinitely in some
environments (interestingly not the jenkins jobs).

To ensure that the correct java is used, we should set JAVA_HOME in the
api_test target.

Change-Id: Ibdead49955f54ad32e00dc477c2a47b236fcbe44
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188145
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>