MB-55960: Temporarily disable history for all collections
QE team has found instability in the elixir branch as kv_engine has not
fully merged forwards changes from neo to master. kv_engine is using the
master branch for Elixir private preview. To ensure stablity for the
private preview we should temporarily disable CDC in Elixir and
re-enable it post-private preview.

On the surface one would expect just not enabling history retention at
the Bucket level would disable CDC, but ns_server defaults history to
true for collections of magma buckets regardless of the history
retention settings on the Bucket, such that the OOTB experience is
better. Setting history to true is the thing that we need to stop doing
as kv_engine cannot cope with it in this partially implemented state as
not all changed have yet to be merged to master.

Change-Id: Idd08739b8fe01295547bbb5c53c96d3a7de5033c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188155
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>