MB-56034: Make a script to generate index workload.
Made a script which creates workload on the indexer by using pillowfight
on all of the collections that exists in a bucket and creating indexes
on Field_1 and Field_2. Since the "Fields" are changing, this means that
the indexes need to be updated, resulting in index workload. Supports
multiprocessing module, allowing for parallelism using the --processes
flag.

python scripts/generate_index_workload.py -b travel-sample

Change-Id: I582bf79fc9d6062fe91d26a59acbbabb5f97c71d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188375
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Dan Ho <dan.ho@couchbase.com>