MB-49059: Undo removal of 'canAbortRebalance'
Undo the removal of a setting 'auto_failover_can_abort_rebalance'
because it turns out it was being used by customers. Make note of the
double-negative -- we are undoing a previous setting removal, to leave
the setting as it was before all of the changes to it.

Also fix api_tests for this change.

Change-Id: I4d4ceb464cd0e6ab5bf0f1f8a4bc0567af32f43f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188303
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>