MB-55615 Return clear error message on ldap failure
Currently LDAP server connection failures return almost no
information to the UI, so the user can't find the reason for failure.
Here we've added message to be received from ELDAP library when the
connection fails. Then we format it and return a more descriptive
error message to the UI. Also empty list of hosts are handled.

Change-Id: Ic1ffb48bc5c7add52d5a4b623e70b81ea24aee45
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188703
Tested-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>