api_tests: Refactor testlib and pass Cluster to BaseTestSet
Makes Cluster a class for future extension and moves it to a module in
a new testlib directory. Also moves testlib.py into this directory.

We now pass the cluster to BaseTestSet, so we can use self.cluster
instead of passing cluster to every function.

Change-Id: I5e135c9ce7f8e6f229206c2a75df801558df5408
Reviewed-on: https://review.couchbase.org/c/ns_server/+/189154
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>