MB-54051 Handle normal vbucket mover exit
When a rebalance is stopped we were expecting any vbucket movers could
be shutdown. There is a race with the process getting shutdown via exit
and the process ending normally. So we now look for "normal" in addition
to "shutdown" as the exit reason.

Change-Id: I9df55a903fd3b5b6ba2c847be1d7527e6596fbc2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188917
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>