api_tests: Add requirement for not connecting all nodes
In order to test adding nodes to the cluster, we may wish to start the
test with extra nodes that aren't yet connected. We do this with a
'num_connected' test set requirement, specifying how many of the
'num_nodes' to connect prior to running the tests.

Change-Id: Id18238b015bc0cc47b8187eeae5d14cbf993f33a
Co-authored-by: Dan Ho <dan.ho@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/ns_server/+/188962
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>