MB-54902: Don't trigger session_expire audit when...
... ui tokens are rotated

When we rotate the ui token it doesn't mean that user session has
expired. The session expires only when all tokens are expired for
that user session.

The idea is the following.
Every time we create a new token for some session, we put a link to
that new token into previous token.

Basically we now maintain a double linked list of tokens
for each session (because each token now has a link to prev and next
tokens).

Then when we are expiring a token we can check if that socket is
the last one in that session by checking the next token field.
If there is no next token, this token is the most recent one.

This patch also fixes the following bug:
If user opens UI page in multiply tabs, tokens "multiply" and
the logout button doesn't always logs out user in all the tabs.
The behavior is random.

Change-Id: I05ddcd28aa8b9a147fbded87a47273c1d08263fd
Reviewed-on: https://review.couchbase.org/c/ns_server/+/189107
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>