api_tests: Refactor ClusterRequirements for parameterisation
Currently we have run.py and testlib.py doing too much requirements
handling. This should be split out into appropriate modules, and the
ClusterRequirements namedtuple should be replaced with a class that
handles creating an appropriate cluster, using classes for each
requirement, so that it is easier to add a new requirement and
understand what the existing requirements do.

These changes make parameterisation simpler to implement, and ensures
that it is practical to extend parameterisation to be more efficient,
by allowing modification of an existing cluster in order to satisfy
requirements.

Change-Id: I02338ff878210dd3aea0fc0988e238909b735626
Reviewed-on: https://review.couchbase.org/c/ns_server/+/189254
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>