Format some sigar stats test data
The current test data is hard to read as it is json written on as few
lines as possible. We can remove the escaped new lines from the json and
just print it in a more standard way to make it easier to read and add
to in the future.

Change-Id: Ie062e2856ab98df8cb947728a60d5d03cd82939c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/189265
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>