MB-57832: [BP] Format some sigar stats test data
The current test data is hard to read as it is json written on as few
lines as possible. We can remove the escaped new lines from the json and
just print it in a more standard way to make it easier to read and add
to in the future.

Change-Id: I3795cf2e93b7917d97e31bb6bd12b83477afce89
Reviewed-on: https://review.couchbase.org/c/ns_server/+/199440
Well-Formed: Restriction Checker
Tested-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>