MB-56411: Add log and configuration file to sigar_port
Change the command line arguments for starting up sigar_port
to:

   sigar_port --babysitter_pid <pid> \
              --logfile <logdir>/sigar_port.log \
              --config <configdir>/sigar_port.json

By default the config file _does not exists_, but it was
added in order to let the user get an easy way to change
the log level for sigar_port iff needed to collect more
data.

A user may create the file with the following content:

    { "loglevel" : "debug" }

loglevel may be set to: "debug", "info" or "error". By default
"error" is used, which means that in a normal deployment one
should barely get anything logged.

sigar_port will use two log files and keep up to 40MB of
historical log data (20MB in the current file, which will
be renamed to .old and start on a new one)

Change-Id: Id5c7e632216aa70e8a633d533ff89fcc7b02a5fa
Reviewed-on: https://review.couchbase.org/c/ns_server/+/189516
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>