MB-24487: Add assertion dupe check
Don't accept the same assertion twice basically.

Note: by default recipient check should guarantee that assertion
can't be reused on another node, but in case if recipient is the
same for all nodes (it can be the case if spCustomBaseURL is used,
or if recipient check is disabled), we still need a protection
against that.

Change-Id: I640608db242cb175a0854b2492f08d72ca3916bf
Reviewed-on: https://review.couchbase.org/c/ns_server/+/189550
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>