MB-55578: Pass MonitorState to health monitor handle_call impl
Currently we pass a number of arguments to the health monitor
implementations as they may need to know node statuses or nodes in the
clusters. In future changes we need to cache additional state in some
monitors and make monitor refresh intervals configurable (with
differing values). Remove the two existing parameters in favour of a
single MonitorState parameter which can:

a) be extended with state common to all/most monitors
b) be extended with state pertinent to only one or two monitors without
   having it be present for all monitors.

Change-Id: I6f2b34889813740101eb347491e1db16ce16607c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/189603
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>