MB-56551: Ensure that samples are loaded in order
We were previously using a stack to track sample loading tasks, but now
that we can potentially receive multiple sample load requests in a row,
we need to use a queue to ensure that the samples are loaded in the
order that the requests came in. If we didn't do this, then a burst of
requests could cause early requests to start after later requests
have completed.

Change-Id: Iad9866433408bee730287759d997e214ff73cd92
Reviewed-on: https://review.couchbase.org/c/ns_server/+/189881
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>