Change backing storage for config_profile
Change the backing storage for config_profile data from application_env
to persistent_term because persistent_term is significantly more
performant for this exact use-case. It was just overlooked originally.

Understanding when to use persistent_term is somewhat tricky but
generally data that is set once, or very infrequently, but accessed a
lot. It essentially stores the data in an unprotected hashmap which
makes it faster than ets tables which require locks.

Change-Id: Ia132c0fadd49e0de41885f5cdff762e7b3021878
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190058
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>