MB-56530: Exit ns_server if config_profile changes
Determine if we are using the same config_profile as the server was
originally started with, on every startup. If the profiles do not match,
we crash the main ns_server process, stopping us from starting the
server. Originally, an error message was thought to be "enough" but the
error really is fatal, and unrecoverable. So instead we are going to
kill the process with an error message.

Change-Id: I64c872bb953414cb1d2c0e39893ad038208a5d76
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190059
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>