Increase node start timeout to avoid api_test job failing
The current timeout is set to 30s, which doesn't appear to be
sufficient, as this sometimes gets reached, causing the api_test job
to fail. Looking at the logs in one of those cases, the node killed was
at a point which was about 3s prior to a successful node being detected
as up. This indicates that the start-up is typically around 30s, rather
than the >30s example being atypically slow.

For now, doubling the timeout to 60s seems a reasonable solution, but
if we see this higher timeout reached, that is probably be a sign that
the server startup time is atypical, so we should not increase the
timeout further.

Change-Id: I8fb71977d13a82630894cc45d53ac468d9dbacc2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190102
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>