Refactor menelaus_web_samples unit test
Unit tests should use a setup fixture for meck setup and teardown, to
ensure that tests do cause each other to fail.
For instance, if meck:new is called for a module that was meant to be
unloaded at the end of another test, but was not, due to test failure,
then that would cause an error.

Change-Id: I1b0f7a287eb3965a1d83454fa01cb6d2adc8491c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190283
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>