MB-56837: Fix TaskTestSet incorrectly checking versions
The original tests for whether the task version changed after a tasks
list change were wrong, but just happened not to fail. In a future
patch, the order of things changes slightly, exposing the mistakes:

- last_task was being set to the task_update rather than a copy of it
- The extras field was being compared to the original task's status,
  which is incorrect, but doesn't break things
- By default, we have a bucket created for the testset, which doesn't
  become warmed up until mid-test.
If things line up correctly, this causes the tasks version to change
when we don't expect it to. This wasn't caught originally as the bucket
will already be warmed up if other tests run at the same time, and when
the test was the first to run, it would become warmed up at a time when
the tasks version was expected to change.
To fix this, all buckets must be deleted during testset setup. It may
be the case that other tasks could cause unexpected tasks version
changes, but as far as I can tell, none of the other tasks that impact
the tasks version should be doing anything unexpectedly if all buckets
are deleted.

Change-Id: I5071db3a236780b14e634b171037410c99f66e2c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190295
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>