MB-24487: Add support for mandatory settings in settings2
Each param that is managed by menelaus_web_settings2 can now be
marked as mandatory. "Mandory" field is a bool or a fun.
If it is a fun it is called with all the existing parameter values
passed to it as an arg. That fun must return true or a false.
By default all params are optional.
For example:
[{"param1", #{cfg_key => k1, type => int}},
 {"param2", #{cfg_key => k2, type => int, mandatory => true}},
 {"param3", #{cfg_key => k3, type => int,
              mandatory => fun (#{k1 := Value}) -> Value == 1 end}}]

In this example:
param1 is optional
param2 is mandatory
param3 is mandatory only if param1 is set to 1

Change-Id: I633862138cb0b546d26e4e5aafe32d906b98be80
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190305
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>