MB-55576: Take into account services when selecting orchestrator node
When deciding whether or not a node should attempt to become the
orchestrator we currently take into account the version of the node and
the name.

For the sake of data avilability it is not ideal to have the
orchestrator run on data service nodes. If the orchestrator becomes
unreachable and it must be failed over then this takes longer than
failing over any other (data) node as the leases supplied by other nodes
to the orchestrator that we are failing over must expire before the new
orchestrator can take any action. Currently this can take up to ~15
seconds. Failover of a data service node can be done in single digit
seconds at the moment.

To improve data availability, prioritise nodes that are not running the
data service for orchestrator placement. Should a customer run a cluster
without a data service node then this node should run the orchestrator
until failed over. This improves data availbility in the event of a
failover provided that there is always a non-data service node present
in the cluster (and that it is the orchestrator).

This is implemented by defining an orchestrator service weight list as
follows:

kv - 10000
index - 1000
fts - 1000
cbas - 1000
n1ql - 100
eventing - 100
backup - 10

Higher weight => lower priority for orchestrator placement. This code
takes into account service-less nodes too as they implicitly have 0
weight (lowest priority). The new priority algorithm will first consider
node version, then services, then finally the node name.

Importantly, it should also be noted that no compat mode change has been
made as one is not necessary here. Consider an older (7.2.X) version
node with the existing orchestrator priority algorithm. It currently
considers version first, then potentially name as a tie-breaker. Newer
versions of Couchbase Server with this new priority algoirthm will be
higher version (e.g. 7.5.0) which will cause the new node to become the
orchestrator with the new priority algorithm. Importantly the higher
version of the newer node means that older nodes will always consider it
to be a higher priority. The new priority algorithm continues to use
version as the first priority criterion. This allows us to tweak the
algorithm in future versions similarly, and have the newer priority
algorithm take effect as soon as the newer version node is introduced
into the cluster.

Also worth noting is the impact of the changes to the
strongly_lower_priority_node() function. The changes to this function
change the criteria under which a node can take over from another as
orchestrator without changing the mechansim by which we do that (i.e. we
take over in the same way, but for new reasons). Notably this now allows
for a node to take over from another of the same version provided that
the services that the current orchestrator are running are lower
priority for orchestrator placement (higher weight) than the candidate
node. This prevents the customer from having to take action to ensure
that the orchestrator is placed on a non-data service node.

Finally, it should be explicitly mentioned that the format of the
heartbeat messages has been kept the same. This required some extra code
to maintain (so it is worth mentioning here). This means that we must
add the services to create the node_info() from the local nodes config.
Services should never change on a node so we shouldn't have to worry
about consistency there. The alternative would have been to have two
different heartbeat formats which would likely have been more code/more
complicated code.

Change-Id: I480308ac5d9f216e95da170bf1ae05d1e030e224
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190509
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>