MB-56463 [BP] Make sure stats are always collected
This is a backport of MB-51620.

When cbcollect uses the _createStatsSnapshot endpoint the results may be
an error message, instead of the expected directory path containing the
stats snapshot. Prior to this change the error message was treated as a
directory and we failed to get stats. With this change we validate the
results to be an existing directory and if it isn't the case then the
stats are obtained directly from disk.

Change-Id: I6fe380af46bc22f7179eeee260d7c9d1ef39d64e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190584
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>