MB-56463 [BP] Return valid error on snapshot creation timeout
This is a backport of MB-51620.

Prior to this change when the _createStatsSnapshot received a timeout
from the snapshot creation it would return the 'timeout' atom which
caused issues in reporting back the error. We now catch this error and
return an appropriate message "Request timed out".

Change-Id: I366b38dd58293a2b6fe13a4410741da48334f26c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190585
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>