Fix cluster tests NumNodes requirement
When we don't specify a number of connected nodes required for a test
set, we wish to have all required nodes connected. If there is a
cluster with an extra node that is not connected, but there are enough
connected nodes to satisfy the num_nodes requirement, then the cluster
should meet that requirement.

Allow test set to run on cluster with enough connected nodes and extra
nodes that aren't connected.

Change-Id: I1815f0285c0af4e438f4223c9ae528008d4e03ea
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190694
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>