MB-56839: Check raw_ram_quota for loading into existing bucket
When loading a new sample bucket, we consider ?SAMPLE_BUCKET_QUOTA to
be a per-node limit, as we multiply it by the number of nodes rather
than directly comparing it to the total cluster quota. We should be
consistent with this when loading into an existing sample bucket.

Change-Id: Idd002f5b2ac69938b271a58ea5cc9e01597249d8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190695
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>