MB-56838: Make memsize requirement strictly equal
We currently have a min_memsize requirement, which means that a testset
may be ran with a higher memsize than specified. It makes more sense
for the requirement to be strict, and costs very little to enforce as
it doesn't require creating any additional clusters.

Change-Id: I5d5102d372b17920d338b2caa28dc6ba6fa9a70e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190748
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>