api_tests: Fix NumNodes requirement string
We merged NumNodes and NumConnected into one requirement, because we
need to check one's value to determine if the other is met in some
cases. When we did this, num_connected was lost from the string
representation. In order to have requirements with multiple arguments
we should keep track of all of the variables of the requirement, and
include them all in the string representation.

Change-Id: Ib710859c9af4f5cb6511a74547d857e4bd7ffa5d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190843
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>