MB-55576: Expand higher_priority_node test in mb_master.erl
We are currently not testing a couple of variations of version and node
name, add those to the test. Rename the test to
higher_priority_node_test() too to make it clearer when we add a
strongly_lower_priority_node_test() in a future change.

Change-Id: I49e355e4fb6f42609969b5961fb227608d708e5d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190981
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>