MB-55047 Add prometheus http service discovery
This adds a new REST api to allow configuring prometheus to use the api
to obtain the scrape targets (as opposed to listing them in the config
file).

The new API is /prometheus_sd_config which accepts query parameters:

        type={json|yaml}
        disposition={inline|attachment}
        port={secure|insecure}
        network={default|external}

The first of the options is the default. When disposition is "attachment"
the extension for filename will be based on "type".

An example of a config file using this endpoint where a user named
ExternalStatsReads with the "external stats reader" role has been added
to the cluster.

scrape_configs:
- job_name: mycmdb
  basic_auth:
    username: Administrator
    password: asdasd
  http_sd_configs:
  - url: http://localhost:9000/prometheus_sd_config?type=json&disposition=attachment
    basic_auth:
        username: ExternalStatsReader
        password: ThanksChris

Prometheus would then be started with the above config file

./prometheus --config.file prom_config --log.level=debug

An example of using the endpoint and having the results go into a file:

$ wget --content-disposition --http-user Administrator --http-password \
  asdasd 'http://127.0.0.1:9000/prometheus_sd_config?disposition=attachment'

Saving to: ‘couchbase_sd_config_CouchbaseTestLab.json’

2023-05-15 16:09:23 (11.0 MB/s) - ‘couchbase_sd_config_CouchbaseTestLab.json’
 saved [69/69]

$ cat couchbase_sd_config_CouchbaseTestLab.json
[{"targets":["127.0.0.1:19000","127.0.0.1:19001","127.0.0.1:19002"]}]

Change-Id: I37b770c55518dec4c599933360d6b72604ce807b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/190994
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>