MB-56470 Unblock ns_audit in case memcached is slow
Connecting to memcached can be slow and become a bottleneck for
ns_audit. Here we're trying to connect to memcached via a
separate process in order to free up ns_audit to do its routine
tasks.

The new process receives one message and tries to send it.
At the end it sends back result to the parent indicating
it can handle more messages or asking for some delay send.
Based on this message, parent decides to send more messages
or pause send for a while.

Change-Id: I05ca1a11299c8b8743eab619cedb4519367824a2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/191240
Tested-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>