cluster_tests: Unspecified requirements should be ignored
Currently an unspecified requirement gets a default value. This can
prevent using an existing cluster for a testset that does not specify
any incompatible requirements. Instead, we should exclude requirements
that have not been specified.

This means that there is a greater range of possible clusters that may
be used for a testset that does not specify all of its requirements.
This increases the non-determinism of the tests, but we make it clear
what cluster the tests were ran on, so insufficient requirements can
be fixed once discovered.

Change-Id: I330c32ff9707d99c4a51a03ba6ebf4bb29c5a663
Reviewed-on: https://review.couchbase.org/c/ns_server/+/191434
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>