cluster_tests: Make connected_nodes requirement strict
The NodeAdditionTests require only 1 node to be connected, so the
requirement should be a strict requirement, instead of >=

When using an existing cluster, if that cluster has more nodes than
the testset requires, this change will make it so that the requirements
are no longer met for that cluster, which is appropriate behaviour for
NodeAdditionTests.
A future change could make it so that we can make modifications to an
existing cluster, such as disconnecting nodes that need to be
disconnected to satisfy the requirements.

Change-Id: I9009148d636c8f156cd1c38725c4c97ae2ab00bd
Reviewed-on: https://review.couchbase.org/c/ns_server/+/191435
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>