MB-57087 "wrap" autogeneratedClientKey so it gets sanitized
The autogeneratedClientKey is wrapped with otpCookie so it gets
sanitized before getting logged. The key must always be wrapped to
handle the case where an older release, even if no longer supported,
attempts to connect to the node. Without the wrapping the older release
will log the unsanitized key.

Change-Id: I44f9a1290a42af3863b9dc6ed6d32e35c3b95a37
Reviewed-on: https://review.couchbase.org/c/ns_server/+/191618
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>