MB-56730: Flag a missing ns_server_monitor update from a node.
When a node stops sending its own monitor statuses to the orchestrator,
we don't flag it as a network partition assuming all other nodes
continue to receive ns_server_monitor heartbeats. The node in question
is assumed to be healthy and the absence of its node_monitor updates is
not flagged as an issue. Flag it as a potential network partition if
the node isn't sending its own monitor statuses to the orchestrator
(maybe because it cannot receive heartbeats from other nodes).

Flagging potential_network_partition doesn't trigger autofailover.
It has no effect other than possibly flagging an alert for
communication issues if the condition persists for 5 minutes.

Change-Id: I3b2e53e0c1ebc85d5f97e90d1dcfde373effd5f4
Reviewed-on: https://review.couchbase.org/c/ns_server/+/191772
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>