MB-57176: Couchdb incorrectly started serverless
Pass the env correctly to CouchDB s/t it won't start in serverless mode.
This was caused when we changed the backing store for the
config_profile. This is associated with a change in CouchDB that stops
using the config_profile module s/t hopefully we don't run into issue
like this again.

Change-Id: I62128ba7c1ceec1f2b3ef2eeaff911b88c8bd548
Reviewed-on: https://review.couchbase.org/c/ns_server/+/191845
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>