MB-55578: Pass MonitorState to health monitor handle_info/cast impl
Currently we pass a number of arguments to the health monitor
implementations as they may need to know node statuses or nodes in the
clusters. In future changes we need to cache additional state in some
monitors and make monitor refresh intervals configurable (with
differing values). Remove the two existing parameters in favour of a
single MonitorState map which can:

a) be extended with state common to all/most monitors
b) be extended with state only pertinent to one or two monitors without
   adding it to other monitors too (which is why we are using a map here
   rather than a record)

Note also that in this change I have moved pattern matching on
parameters within the MonitorState from the function head to the
function body. Whilst it is possible to match the parameters in the
function head, it is error prone here as the monitors have handle_cast
and handle_call function clauses that match anything and simply log an
error. Rather than pattern match in the function head, we should be more
permissive with our attempts at pattern matching (i.e. only pattern
match on the message that we are attempting to process) and then
retrieve the keys from the MonitorState map separately which should
result in a pattern matching error rather than ignoring (logging) it.

Change-Id: I124239d1a9bad70dac0823d5907f528dcad4a2a5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/189585
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>