MB-57162: Accept message types from pre-elixir service_agents ...
... in service_manager.

When a service rebalance is run we pick a leader node based on the
priority set for each node the service is hosted on. Services (like
eventing in this case) can choose to give higher priority to a lower
version node and therefore the leader could be running old service_agent
code. The service_manager though will always run on the orchestrator
node in ns_server and hence it should be backwards compatible and accept
the old message types from pre-elixir service_agents.

Change-Id: I556d4f714ce32ba46445cf8bf022601292009e55
Reviewed-on: https://review.couchbase.org/c/ns_server/+/191928
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>