MB-56470 Continue sending logs if einval error happens
Currently when an einval error happens, we stop sending audit
logs to memcached. The einval case is a programming error.
We're audit logging something which memcached considers a
"hard" failure and will never succeed. In this case we should
continue sending the rest of queue to memcached. This change
will fix the above issue.

Change-Id: I6d4dd260aea38cf8010a207c92f0a74ffb51fbfa
Reviewed-on: https://review.couchbase.org/c/ns_server/+/191943
Tested-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>