MB-55508: Split config and memcached steps in janitor cleanup
This change splits the janitor run for a bucket into two phases. The
first phase does any pontential bucket config updates, and the config
from that gets passed to second phase that does interactions with
memcached. The intention is to make it easier to batch a set of buckets
to the janitor for cleanup.

Change-Id: I7f083d24a33d1a580a6f03c61a96714291bf5989
Reviewed-on: https://review.couchbase.org/c/ns_server/+/192965
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>