MB-55579: Bucket storage mode migration.
Bucket storage mode can be now migrated via a series of swap rebalance
or failover + full recovery approach.

When the storage mode for a bucket is updated in chronicle, we add
per-node override storage_mode props for all the nodes the bucket is
hosted on.

Eventually, as and when old nodes are rebalanced-out (and new nodes are
added in), we remove the per-node override storage_mode props of the old
nodes. At any point during a swap rebalance the bucket config could look
as below:

[{storage_mode, magma}, {servers, [n1, n2]},
 {{node, n1, storage_mode}, couchstore}]

The way to interpret this is: If a override prop for a node isn't
present, it's storage_mode is the global storage_mode, else it's the
per-node override storage_mode. In the above example the storage mode on
node n1 is couchstore and on node n2 is magma.

When a node is failed-over we don't clear per-node storage_mode from all
the bucket, but it's removed when the node recovery_type is set to full
and is rebalanced in.

The per-node storage mode is present in the
/pools/default/buckets/<bucket-name> and
/pools/default/bucketsStreaming/<bucket-name> endpoints.

For example:

curl -u Administrator:asdasd
http://127.0.0.1:9000/pools/default/buckets/foo | jq '.storageBackend'
"magma"

curl -u Administrator:asdasd
http://127.0.0.1:9000/pools/default/buckets/foo | jq
'.nodes[].storageBackend' "couchstore" "couchstore"

In addition to the storage_mode key we also add override props for the
autocompaction key (which when set to non-default values can be
different for couchstore and magma) during the storage_mode migration.

Change-Id: I3c26ac25fed5aead917344118a0ff8d6573524ea
Reviewed-on: https://review.couchbase.org/c/ns_server/+/192124
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>