cluster_tests: Wait for rebalance when create/delete bucket
When the number of nodes requirement wasn't specified, we default to 1
node, and cluster_run_lib still rebalances by default, but doesn't wait
for the rebalance to finish. There are a few options here:
- wait for the rebalance to finish in all cases
- don't rebalance by default
- ensure that tests do not assume that the cluster won't be rebalancing
  when they start

The main reason that tests rely on the cluster not being in a
rebalancing state is that they need to create or delete buckets.
This change adds methods to Cluster for making these requests, which
additionally wait for any rebalance to complete. In the unlikely case
where we wish to test creating/deleting buckets during a rebalance,
we can use the existing testlib.post/delete/etc functions, or add a
parameter to disable waiting for rebalance.

Change-Id: If9906e03a310603daf3d67781cf1715299a952f5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/192074
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>