Allow rebalance plot to show rebalance with unfinished moves
When a rebalance is stuck, moves will have null times, which causes
our plot-bucket rebalance script to crash. We should handle these by
giving them a max value (so the bars go to the end of the graph).
When this happens, it would not be obvious which vbuckets are stuck, so
I've added a red border for the vbuckets with null times.

Change-Id: I4f30f06f2f3b63e8afa4f16a5b3bd054e36493f0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/192255
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>