MB-59364: [BP] Bulk config updates over buckets for failover
MB-55508:
This change organizes the failover in 3 steps:
1) Prep Phase - A phase that just prepares the failover parameters for
buckets. This change is stateless and does not affect any state of the
system in config or anywhere else.

2) Bulk config update - This updates the config over a number of buckets
for the failover using a single transaction.

3) Janitor run - This runs the janitor over the buckets(one at a time)
for which config was set in step 2.

Steps 1 is run over all buckets. The results of step 1 are fed into the
other steps. Steps 2 and 3 can be run over a group of buckets or even
one bucket at a time like the original way.

The internalSettings parameter "failoverBulkConfigFactor" can be used
to control how aggressive we want to be in doing the bulk updates.

Change-Id: I61989014901cf49892ab9616562e703448230a03
Reviewed-on: https://review.couchbase.org/c/ns_server/+/192323
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201654
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker