MB-56842: Add serviceless node test to cluster_test suite
This adds a test to the cluster_tests which adds nodes without any
services to a cluster, does a failover of a node, and then recovers and
adds back the node. The orchestrator node is removed and then we check
that a new orchestrator (also serviceless) has been choosen.

Then the same is done for serviceless nodes joined to a cluster.

Change-Id: Iaeb3ea8ef76c95eaae556355e09611c69991e67f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/192278
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>