MB-55508: Janitor a set of buckets at a time during failover
This hooks failover into the multi bucket janitor function. The janitor
internally runs memcached interactions in parallel over these buckets.
We decide on how many buckets to janitor at a time based on a default
and config override as well as the number of schedulers online. The
internal settings parameter "failoverBulkJanitorFactor" can be used to
set a factor to tune how aggressive we want to be with parallelization.
Internally we always cap the parallelization with respect to the number
of schedulers online to the system as we do not try to parallelize over
buckets more than the number of cores available to the system.

Change-Id: I54941395e638fc1ccb396605c6efc3cb037c94dd
Reviewed-on: https://review.couchbase.org/c/ns_server/+/192968
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>