MB-57639: Check collection manifest has changed before updating
When a request to modify the collection is sent with the existing
collection manifest, the order of properties may change, causing the
manifest to be treated as different, incrementing the id, as well as
modifying the manifest. We should ensure that the manifest only changes
if the props actually change.

Change-Id: I45259d3cbc922ed822796fc55d71bbeac0550923
Reviewed-on: https://review.couchbase.org/c/ns_server/+/193168
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>