[cluster_tests] Ability to specify seed for tests
Add option "--seed <string>", which sets the random generator seed.

If not set, random seed is used (python uses time base seed, but we
generate random string, and use this string as a seed in order to
make it reusable).

In the end of all tests it outputs the seed, which can be
copy-pasted to the "--seed" arg.

Few points:
- All setup's are called with the same seed (so testsets don't affect
  each other, behavior should not depend on how many testsets are
  run);
- All tests inside testset are executed with the same seed (so tests
  do not affect each other, behavior should not depend on how many
  tests are run);
- All teardown's are executed with the same seed (behavior should
  not depend on how many tests are run);
- Setup seed, test seed, and teardown seed are not the same but they
  are derived from the "main seed".

Change-Id: I6d3ba42806324acb495cad7eba9a2b608e33ef3a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/193614
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>