MB-55578: Make index_monitor implement health_monitor behaviour
This will allow us to modify it's refresh period in the same way as the
other monitors can be modified.

Note that this reduce the refresh inteval from 2s to 1s (the default for
the health monitor behaviour) for the sake of consistency and
simplicity.

Change-Id: If856b3ad279a5a02bcfb43c89c7c0f2c7ae49bf0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/193991
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: <varun.velamuri@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>